HMAN-985: Include HMI endpoint health in outbound adaptor health
Added HearingManagementInterfaceHealthIndicator.  This adds a check of the Hearing Management Interface private health endpoint to the outbound adapater health output.

The outbound adapter health will be shown as down if the Hearing Management Interface private endpoint returns an error, or an error is encountered when attempting to connect to it.  The outbound adapter health will also be shown as down if the call to Active Directory to get an access token returns an error, or an error is encountered when attempting to connect to it.

A retryer has been added to retry the connection to Active Directory.  This is configurable via environment variables.

Cleaned up some tests and changed logging to use placeholders instead of string concatenation.